requirements.txt
fastapi
uvicorn
python-jose[cryptography]
httpx
dotenv
